% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolImport.R
\name{metabol.import}
\alias{metabol.import}
\title{Import metabolomics data and metadata}
\usage{
metabol.import(
  raw_data,
  metadata = NULL,
  strict = TRUE,
  assign_result = TRUE,
  assign_name = "imp_data",
  envir = parent.frame()
)
}
\arguments{
\item{raw_data}{A data.frame or tibble containing a 'Sample' column and metabolite quantifications in the remaining columns.}

\item{metadata}{Optional data.frame or tibble containing at least a 'Sample' column. A 'Group' column is recommended.}

\item{strict}{Logical; if TRUE, stops execution on data validation errors. Default is TRUE.}

\item{assign_result}{Logical; if TRUE, assigns the object to the environment. Default is FALSE.}

\item{assign_name}{Character; name to assign the object if assign_result = TRUE.}

\item{envir}{Environment in which to assign the object if assign_result = TRUE. Default is parent.frame().}
}
\value{
An object of class \code{metabolR} containing data, metadata, warnings, and summary counts.
}
\description{
Prepares raw metabolomics data and optional metadata into a standardized \code{metabolR} object.
Ensures consistent formatting, validation, and easy downstream processing.
}
\examples{
raw_data <- data.frame(
  Sample = c("S1", "S2", "S3"),
  Met1 = c(10, 15, 12),
  Met2 = c(5, 3, 6)
)
metadata <- data.frame(
  Sample = c("S1", "S2", "S3"),
  Group  = c("Control", "Treatment", "Control")
)

imp <- metabol.import(raw_data, metadata)
print(imp)

}
\references{
Xia J, Sinelnikov IV, Han B, Wishart DS (2015). MetaboAnalyst 3.0—making metabolomics more meaningful. \emph{Nucleic Acids Research}, 43(W1): W251–W257. \url{doi:10.1093/nar/gkv380}
Pang Z, Chong J, Zhou G, de Lima Morais DA, Chang L, Barrette M, et al. (2021). MetaboAnalyst 5.0: narrowing the gap between raw spectra and functional insights. \emph{Nucleic Acids Research}, 49(W1): W388–W396. \url{doi:10.1093/nar/gkab382}
}
