% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolQC.R
\name{metabol.qc}
\alias{metabol.qc}
\title{Metabolomics Quality Control (QC) Plots and Summary}
\usage{
metabol.qc(
  normalized_data,
  metadata,
  group_col = "Group",
  cluster_rows = FALSE,
  cluster_cols = FALSE,
  angle_col = 45,
  display_corr_values = FALSE
)
}
\arguments{
\item{normalized_data}{A \code{metabolNorm} object or list containing an \code{expr_matrix}.}

\item{metadata}{Data.frame containing at least 'Sample' column and grouping column.}

\item{group_col}{Character. Column name in \code{metadata} specifying the group/factor. Default "Group".}

\item{cluster_rows}{Logical. Whether to cluster rows in correlation heatmap. Default FALSE.}

\item{cluster_cols}{Logical. Whether to cluster columns in correlation heatmap. Default FALSE.}

\item{angle_col}{Numeric. Angle for x-axis labels. Default 45.}

\item{display_corr_values}{Logical. Whether to display correlation values in heatmap. Default FALSE.}
}
\value{
Invisibly returns a list containing:
\item{data_ready}{Normalized expression matrix used for QC.}
\item{presence_per_group}{Presence/absence matrix per metabolite per group.}
\item{corr_matrix_samples}{Pearson correlation matrix across samples.}
\item{plot_density}{Density plot of concentrations per group.}
\item{plot_boxplot}{Boxplot of concentrations per sample.}
\item{histograms_group}{List of mean ± SD barplots per metabolite per group.}
\item{summary}{QC summary statistics previously printed to console.}
}
\description{
Generates QC plots and summaries for normalized metabolomics data,
including density plots, boxplots, correlation heatmaps, and mean ± SD per metabolite per group.
}
\examples{
\dontrun{
# Minimal example
set.seed(123)
raw_data <- data.frame(
  Sample = paste0("S", 1:10),
  Met1 = rnorm(10),
  Met2 = rnorm(10)
)

meta <- data.frame(
  Sample = paste0("S", 1:10),
  Group = rep(c("A","B"), each = 5)
)

imp_data <- list(data = raw_data, metadata = meta)
class(imp_data) <- "metabolR"

norm_obj <- metabol.normalize(imp_data)

metabol.qc(norm_obj, metadata = meta)
}

}
\references{
Dunn WB, Broadhurst D, Begley P, et al. (2011). Procedures for large-scale metabolic profiling of serum and plasma using gas chromatography and liquid chromatography coupled to mass spectrometry. \emph{Nature Protocols}, 6(7): 1060–1083. \url{doi:10.1038/nprot.2011.335}
Broadhurst D, Goodacre R, Reinke SN, et al. (2018). Guidelines and considerations for the use of system suitability and quality control samples in mass spectrometry assays applied in untargeted clinical metabolomic studies. \emph{Metabolomics}, 14: 72. \url{doi:10.1007/s11306-018-1367-3}
}
