% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolTtest.R
\name{metabol.ttest}
\alias{metabol.ttest}
\title{T-test or Mann-Whitney Test for a Metabolite}
\usage{
metabol.ttest(
  normalized_data,
  metadata,
  metabolite,
  group_col = "Group",
  utest = FALSE,
  return_type = c("htest", "all")
)
}
\arguments{
\item{normalized_data}{List returned by \code{metabol.normalize()}, containing \code{$expr_matrix}.}

\item{metadata}{Data.frame containing at least columns 'Sample' and the grouping column.}

\item{metabolite}{Character. Name of the metabolite to test.}

\item{group_col}{Character. Column in \code{metadata} defining groups (default = "Group").}

\item{utest}{Logical. If TRUE, performs Mann-Whitney instead of t-test (default FALSE).}

\item{return_type}{Character. "htest" (returns only test object) or "all" (returns test, data, and plot). Default "htest".}
}
\value{
Invisibly returns a list or htest object depending on \code{return_type}.
}
\description{
Performs a t-test (or Mann-Whitney test) comparing a single metabolite between two groups.
}
\examples{
raw_data <- data.frame(Sample = paste0("S", 1:6),
                       Lactate = rnorm(6))
meta <- data.frame(Sample = paste0("S", 1:6), Group = c("A","A","B","B","A","B"))
imp_data <- list(expr_matrix = raw_data)

metabol.ttest(imp_data, metadata = meta, metabolite = "Lactate", return_type = "all")

}
\references{
Student. (1908). The probable error of a mean. \emph{Biometrika}, 6(1): 1–25. \url{doi:10.2307/2331554}
Mann, H.B., & Whitney, D.R. (1947). On a test of whether one of two random variables is stochastically larger than the other. \emph{Annals of Mathematical Statistics}, 18: 50–60. \url{doi:10.1214/aoms/1177730491}
}
