% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolVenn.R
\name{metabol.venn}
\alias{metabol.venn}
\title{Venn Diagram for Metabolite Presence Across Groups}
\usage{
metabol.venn(normalized_data, cutoff = 0, results = FALSE)
}
\arguments{
\item{normalized_data}{Object returned by \code{metabol.qc()}, containing \code{$presence_per_group}.}

\item{cutoff}{Numeric. Minimum value to consider a metabolite as present (default = 0).}

\item{results}{Logical. If TRUE, returns detailed lists of metabolites; default FALSE.}
}
\value{
Invisibly returns a list with metabolites by group, unique metabolites, and shared metabolites (if results = TRUE). Also prints a Venn diagram.
}
\description{
Generates a Venn diagram showing shared and unique metabolites across groups.
}
\examples{
\dontrun{
presence_per_group <- data.frame(
  MetaboliteID = paste0("Metab", 1:7),
  GroupA = c(1,1,0,2,2,0,1),
  GroupB = c(0,1,3,1,0,1,3)
)

qc_res <- list(presence_per_group = presence_per_group)

metabol.venn(qc_res, cutoff = 0, results = TRUE)
}

}
\references{
Venn, J. (1880). On the diagrammatic and mechanical representation of propositions and reasoning. \emph{The London, Edinburgh, and Dublin Philosophical Magazine and Journal of Science}, 10: 1–18. \url{doi:10.1080/14786448008626994}
Chen, H., & Boutros, P.C. (2011). VennDiagram: a package for the generation of highly-customizable Venn and Euler diagrams in R. \emph{BMC Bioinformatics}, 12: 35. \url{doi:10.1186/1471-2105-12-35}
}
