% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.metabolNorm.R
\name{print.metabolNorm}
\alias{print.metabolNorm}
\title{Print Method for metabolNorm Objects}
\usage{
\method{print}{metabolNorm}(x, ...)
}
\arguments{
\item{x}{An object of class \code{metabolNorm}.}

\item{...}{Additional arguments (ignored).}
}
\description{
Custom print method for objects of class \code{metabolNorm}.
}
