% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MetabolDimred.R
\name{metabol.dimred}
\alias{metabol.dimred}
\title{Dimensionality Reduction for Metabolomics Data}
\usage{
metabol.dimred(
  normalized_data,
  metadata,
  n_neighbors = NULL,
  ncomp_plsda = 2,
  cross_val = 5,
  verbose = TRUE,
  assign_result = TRUE,
  assign_name = "dimred_data",
  envir = parent.frame()
)
}
\arguments{
\item{normalized_data}{A \code{metabolNorm} object or a numeric data.frame/matrix.
Rows are samples, columns are metabolites.}

\item{metadata}{A data.frame containing at least a 'Sample' column and a group/condition column.}

\item{n_neighbors}{Integer, number of neighbors for UMAP. Defaults to half of the number of samples if NULL.}

\item{ncomp_plsda}{Integer, number of PLS-DA components to compute. Default 2.}

\item{cross_val}{Integer, number of folds for PLS-DA cross-validation. Default 5.}

\item{verbose}{Logical, whether to show messages and plots. Default TRUE.}

\item{assign_result}{Logical, whether to assign the result to the environment. Default = TRUE.}

\item{assign_name}{Character, name of the object if assigned. Default = "vip_data".}

\item{envir}{Environment where the object will be assigned. Default = parent.frame().}
}
\value{
A list with the following elements (invisible):
\describe{
\item{expr_matrix}{The numeric matrix used for dimensionality reduction.}
\item{metadata}{Metadata used in the analysis.}
\item{res_dimred}{A list containing \code{plsda_res}, \code{pca_res}, \code{umap_res}, \code{vip_scores}.}
}
}
\description{
Performs PCA, UMAP, and PLS-DA on normalized metabolomics data.
Returns plots and a list with results including VIP scores for PLS-DA.
}
\examples{
\dontrun{
# Minimal example
set.seed(123)
expr_mat <- matrix(rnorm(100), nrow = 20, ncol = 5)
rownames(expr_mat) <- paste0("S", 1:20)
colnames(expr_mat) <- paste0("Met", 1:5)

meta$Sample <- factor(meta$Sample)
meta <- data.frame(
 Sample = paste0("S", 1:20),
 Group  = rep(c("A","B"), each = 10)
)

res <- metabol.dimred(expr_mat, metadata = meta, cross_val = 3)
}
}
\references{
Jolliffe IT (2002). Principal Component Analysis. 2nd edition. Springer Series in Statistics.
Barker M, Rayens W (2003). Partial least squares for discrimination. \emph{Journal of Chemometrics}, 17:166–173. \url{doi:10.1002/cem.785}
McInnes L, Healy J, Melville J (2018). UMAP: Uniform Manifold Approximation and Projection for Dimension Reduction. \emph{arXiv:1802.03426}.
}
